refactor subscribe and add deviceId and verification Token:
